+++ src/main/java/com/example/electriccomponentsshop/dto/SkuDTO.java
@@ -5,11 +5,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class SkuDTO {
     private String id;
-    private Integer quantity;
+    private BigInteger quantity;
 }
