+++ src/main/java/com/example/electriccomponentsshop/entities/CartItem.java
@@ -6,6 +6,9 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @Entity
@@ -20,9 +23,9 @@ public class CartItem {
     @JoinColumn(name="product_id")
     private Product product;
 
-    private Integer quantity;
+    private BigInteger quantity;
 
-    private Double subTotal;
+    private BigDecimal subTotal;
 
     @ManyToOne(fetch =FetchType.LAZY)
     @MapsId("cartId")
