+++ src/main/java/com/example/electriccomponentsshop/entities/ExportItem.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @Getter
 @Setter
@@ -16,7 +17,7 @@ import javax.persistence.*;
 public class ExportItem {
     @EmbeddedId
     private ExportItemId exportItemId;
-    private Integer quantity;
+    private BigInteger quantity;
     @ManyToOne(fetch = FetchType.EAGER)
     @MapsId("skuId")
     @JoinColumn(name = "sku_id", referencedColumnName = "id")
